requirements.txt
Django==3.2.19
django-filter==23.2
psycopg2==2.9.1
psycopg2-binary==2.9.1
django-model-utils==4.3.1
djangorestframework==3.12.4
Markdown==3.2.1
django-cors-headers==3.2.1
coreapi==2.3.3
django-translation-flags==1.0.6
Pillow==9.5.0
sendgrid==6.10.0
django-colorfield==0.9.0
google-cloud-storage==1.32.0
oauth2client==4.1.3
woocommerce==2.1.1
python-magic==0.4.22
django-ckeditor==6.0.0
xlrd==2.0.1
python-slugify==8.0.1
progress==1.6
django-import-export==3.2.0
gunicorn==20.1.0
